Updated the scanner and made it faster.
