+++ Big Neon/Screens/Explore/TicketScannerViewController/ScannerViewModel.swift
